50 half move rule added. Saw a seg fault once running random... May need to sort this before Monte_Carlo
M all_moves.cpp
M check_check.h
M chess.cpp
M convert2bits.cpp
M convert2board.cpp
M game_over.cpp
M global.h
M initialize.h
M piece_moves.h
M preform_move.h

+++ global.h
@@ -28,7 +28,7 @@ struct Chess_Board {
 
 	uint64_t All_Pieces;
 
-	uint16_t Parameters;
+	uint32_t Parameters;
 };
 extern Chess_Board bitboards;
 
@@ -39,7 +39,7 @@ struct Moves_s {
 	//Move_Tree *parents;
 
 	uint64_t bitmove;
-	uint16_t parameters;
+	uint32_t parameters;
 	int piece;
 };
 struct Move_Tree_s {
